feat(client): implement Attachment API
